Modify delay function

+++ OLED_Software_IIC/OLED/oled.h
@@ -27,7 +27,7 @@
 				   
 
 //OLED控制用函数
-void delay_ms(unsigned int ms);
+void delay_ms(unsigned long ms);
 void OLED_ColorTurn(uint8_t i);
 void OLED_DisplayTurn(uint8_t i);
 void OLED_WR_Byte(uint8_t dat,uint8_t cmd);
